Refactor server transformations.
